Implemented RecyclerView with ListAdapter, added onLike and onShare listeners

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,11 +1,11 @@
 package ru.netology.nmedia.dto
 
 data class Post(
-    val id: Int,
+    val id: Long,
     val author: String,
     val published: String,
     val content: String,
-    val likes: Int = 0,
-    val shares: Int = 0,
+    val likes: Long = 0,
+    val shares: Long = 0,
     val likeByMe: Boolean = false
 )
